requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-location-field==2.7.2
django-phonenumber-field==7.2.0
six==1.16.0
Babel==2.13.1
Pillow==10.1.0